# Dynamically initialize grid storage