--- a/tests/ref/grid_reset/size.json
+++ b/tests/ref/grid_reset/size.json
@@ -1,1 +1,1 @@
-{"width":939.0,"height":1020.0,"cell_width":8.0,"cell_height":18.0,"padding_x":2.0,"padding_y":2.0}
+{"width":2532.0,"height":1380.0,"cell_width":9.0,"cell_height":19.0,"padding_x":3.0,"padding_y":3.0}