… tests/ref/grid_reset/alacritty.recording
… tests/ref/grid_reset/grid.json