@jedahan 2018-09-24
Also fixes #977